added adaptive layout to part of app

+++ app/Http/Controllers/Controller.php
@@ -44,7 +44,7 @@ class Controller extends BaseController
             $name = 'image_' . uniqid()  . '.' . end($photo);
             $path = public_path() . '\photo\\';
             move_uploaded_file($request['photo'], $path . $name);
-            if (file_exists('photo\\'.$user->photo))
+            if (file_exists('photo\\'.$user->photo) && $user->photo != null)
             {
                 unlink('photo\\'.$user->photo);
             }
